specify use of the utf8 I/O layer
Solves problem where invoking a method via osrf_http_translator
resulted in an exception thrown by FreezeThaw if any of the characters
in the message are non-ASCII *and* not in the Latin-1 character
set.  Oddly, if Evergreen been originally developed in Georgia the
country, not Georgia the state, the bug would have been apparent much earlier.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>

git-svn-id: svn://svn.open-ils.org/OpenSRF/branches/rel_1_6@2211 9efc2488-bf62-4759-914b-345cdb29e865