Add DOH-based unit test harness for OpenSRF JavaScript
The Dojo Objective Harness enables us to run unit tests from the command
line or within a browser. Running tests from a command line makes it easier
to fold into the continuous integration server. Included in this commit
is a small README for setting up command-line testing and a sample set
of unit tests that exercise parts of JSON_v1.js.


git-svn-id: svn://svn.open-ils.org/OpenSRF/branches/rel_2_0@2223 9efc2488-bf62-4759-914b-345cdb29e865