Make javascript unit tests more granular to ease tracking down errors
Noticed a failure when running with Rhino 1.7RC2 that wasn't happening
with version of Rhino packaged in shrinksafe.jar, and it was a bit
tedious narrowing down which test was actually failing. Breaking up
the tests into smaller sets will make this process easier and might invite
the submission of other tests to target areas that are not yet covered.


git-svn-id: svn://svn.open-ils.org/OpenSRF/branches/rel_2_0@2225 9efc2488-bf62-4759-914b-345cdb29e865