Handle the common NXDOMAIN problem with Python a bit more gracefully
Rather than dumping a nasty full stacktrace to the command line, we
print a hopefully helpful pointer to the actual problem in plain
English.

Also, rather than stopping everything if a "stop_all" command was
issued, skip the error status in osrf_ctl.sh for the Python bit
and carry on shutting down the other services. A bit more convenient
if you want to have Python enabled but don't necessarily need it
running.


git-svn-id: svn://svn.open-ils.org/OpenSRF/branches/rel_2_0@2227 9efc2488-bf62-4759-914b-345cdb29e865