Specify explicit branches for each builder, lazy-style
SVNPoller was able to rely on change filters to look at every
commit to the repository and hand them off to the scheduler
accordingly; GitPoller, however, looks at one branch per
poller, so we need to instantiate one poller per branch of
interest. This gives us an opportunity to refactor our approach
to be more programmatic.

Signed-off-by: Dan Scott <dan@coffeecode.net>