Turn on email and IRC notification about build breaks
Buildbot can notify us when a build breaks via email and IRC.

Now that everything is green on testing.evergreen-ils.org, let's
turn that capability on. As configured, we should only be notified
when a build goes from success to failure, the first time; and
then again, when a build goes from failure to success.

Email notification should go to "interested users" - all of the
email addresses in the collected signed-off-by / author tags in
the commits for a given build.

Signed-off-by: Dan Scott <dan@coffeecode.net>