Move the buildbot workdirs to a non-volatile directory
/tmp/ is a good directory for throwing temporary things,
but not so good for things you want to keep around for a relatively
long time. Especially on Ubuntu, which has a default policy of
deleting files in /tmp/ on reboot - meaning that the git repos
that form the basis of the repository have all of their files
wiped - greatly confusing the buildbot, which tries to run
'git reset --head' in an empty directory and fails miserably
in that situation.

So - ensure that WORKDIR is set to something that exists and
which will exist for the duration of your CI server, and you'll
be all right.

Signed-off-by: Dan Scott <dan@coffeecode.net>