Add SIGPIPE retry handling to child data sysread
Similar to the SIGPIPE retry logic wrapped around the parent process'
syswrite call (for sending data to a child process), protect the child's
sysread call (as it reads data from the parent).  In pre-2.0, the
sysread step was handled by Net::Server, but now we need to protect it
ourselves.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>