Add explicit chown command to README
Warren Layton noticed that there was no explicit command to change the
ownership of the files in the /<PREFIX> directory to be owned by
"opensrf", although there was a statement that the files needed to be
owned by "opensrf". To reduce the chance of failure, add the explicit
command.

TODO (for a willing volunteer): teach the installer to change the
ownership at the time the files are installed!

Signed-off-by: Dan Scott <dscott@laurentian.ca>