OpenSRF Validator Service
Add a new Validator service, and EmailAddress validators.

The service runs a chain of one or more validators, each one being fed the
normalized output of the previous one.

The return from each validator should be a hash of valid (0 or 1), the new
normalized output (the untouched input if invalid or nothing needed to be
changed), and if invalid an error string. Optionally, a validator can also
include an "additionals" hash of extra information to be included in the
final response.

The complete list of validators included is:

OpenSRF::Application::Validator::Base
    The base validator. Always returns valid.
OpenSRF::Application::Validator::Invalid
    Always returns invalid for testing purposes.
OpenSRF::Application::Validator::EmailAddress::Regex
    Does a very basic regular expression check on email addresses.
OpenSRF::Application::Validator::EmailAddress::DNS
    Uses Net::DNS to look up the domain on an email address

Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>
Signed-off-by: Dan Scott <dscott@laurentian.ca>