Remove reference to Fedora specific version in prereq installer
Fedora changes every 6 months, so it's probably silly to have a make
target of "fedora14" when that is no longer supported by the Fedora
project itself. As Fedora has packaged all of the OpenSRF dependencies
(thanks, Ben Webb!), in theory it should be supported by the latest
releases of Fedora in the future... as long as we don't introduce any
dependencies on deprecated versions of packages.

Signed-off-by: Dan Scott <dscott@laurentian.ca>