Add "tar" to base prereqs to appease autotools
On a minimal install of Fedora 17, I was running into the perplexing
error message whilst running 'autoreconf -i':

libtoolize: can not copy `/usr/share/libtool/config/ltmain.sh' to `./'

This was confusing because a manual copy would, in fact, copy the file
just fine.

http://lists.gnu.org/archive/html/libtool/2009-07/msg00030.html finally
lead to enlightenment: autotools uses "tar" to copy files, not "cp".
Thus, to avoid similar head-scratching install problems in the future,
add "tar" to the base prerequisites for building OpenSRF (both to the
DEBs list and the RPMs list).

Signed-off-by: Dan Scott <dscott@laurentian.ca>