Log redaction for sensitive input values, Perl side
This commit attempts to do the same as the C log redaction fix,
but now at the Perl level.  The Perl configuration code was a
little more crufty than the C side, so an additional feature was
added to Config.pm to support the new 'shared' section.  At some
point we should consider a ground-up rewrite of Config.pm, as the
code seems to suffer some from its INI file roots.

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Dan Scott <dan@coffeecode.net>