Enable client logtrace with environment vars
Supports a new environment variable OSRF_LOG_CLIENT which, if true,
enables control and generation of the client log trace value.  This is
the same as setting <client>true</client> within the opensrf core
configuration file.

As a shortcut, if the MOD_PERL environment variable is set, assume
client=true.

This allows clients and non-clients to share an opensrf core
configuration file, when previously the only difference between the two
was the <client> setting.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>