Remove the reference to Evergreen in the README
Stephen Wills reported problems installing Evergreen because of missing
dependencies, and reading the OpenSRF README shows that it refers to the
Evergreen prerequisite installer in a confusing way that might lead to
users using only the Evergreen Makefile.install, rather than using the
OpenSRF Makefile.install.

In addition, Fedora 16 is going to be out of support soon, and Fedora 17
and 18 are well-tested targets now.

Signed-off-by: Dan Scott <dscott@laurentian.ca>