LP#1188195: Default per-process client locale (Perl)
Get/set the locale used by all new client sessions
for the current process.  This is primarily useful
for clients that wish to make a series of opensrf
calls and don't wish to set the locale for each new
AppSession object.

The caller should reset the locale when done using
reset_locale(), as the locale will otherwise persist
for the current process until set/reset again.

This is not for SERVER processes, since they
adopt the locale of their respective callers.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>