Reorder changes to ejabberd.cfg
I found it annoying that the list of changes to make to ejabberd.cfg didn't follow the order that the options showed up in the default Debian ejabberd.cfg.  I reordered them so after you finish changing one option, you can search forward in the document for the next term.