LP#1702978: memcache Get methods use key as va_list format
And, when a key (composed of, say, a username or barcode) has a % in it,
bad things happen.  We will stop acting as if these are variadic functions
now, and also update Evergreen so that it does not do that either.

TODO: Make these actually non-variadic, but that breaks ABI.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Cesar Velez <cesar.velez@equinoxinitiative.org>
Signed-off-by: Jason Stephenson <jason@sigio.com>