LP#1776510 JS libs handle transport errors
Teach the websocket client code to look for the transport_error flag
applied to the websocket wrapper message by the websocket gateway when a
request for an unavilable service is made.

When encountered, fire the transport or generic error handler callbacks,
if available.  Avoid any attempts to further process the message.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>